taskmanager/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
pytest==8.1.1
pytest-django==4.8.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2==2.9.9
django-filter == 24.2
model-bakery==1.17.0